Update InteractionRequest to be used easely with InteractionProviders
This commit changes the InteractionRequest in a way that the constructor is not available anymore.
Advantage of that is that when .Register() is used, an event is raised, and its event args contain
the newly added InteractionRequest.
The big advantage is that very little code is required to add a new InteractionRequest at any time
and it will be automatically registered to the InteractionProvider by the bootstrapper.
This allows efficient and clean code to create UserInteractions.